+++ astro.config.mjs
@@ -1,0 +1,9 @@
+import { defineConfig } from 'astro/config';
+
+// https://astro.build/config
+import react from "@astrojs/react";
+
+// https://astro.build/config
+export default defineConfig({
+  integrations: [react()]
+});
