--- src/App.js
@@ -1,37 +1,0 @@
-import React, { useState, useEffect } from "react";
-import Preloader from "../src/components/Pre";
-
-import Home from "./components/Home";
-
-import {
-  BrowserRouter as Router,
-  Route,
-  Routes,
-} from "react-router-dom";
-import "./style.css";
-import "./App.css";
-
-function App() {
-  const [load, upadateLoad] = useState(true);
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      upadateLoad(false);
-    }, 1200);
-
-    return () => clearTimeout(timer);
-  }, []);
-
-  return (
-    <Router>
-      <Preloader load={load} />
-      <div className="App" id={load ? "no-scroll" : "scroll"}>
-        <Routes basename={process.env.PUBLIC_URL}>
-          <Route path="/" element={<Home />} />
-        </Routes>
-      </div>
-    </Router>
-  );
-}
-
-export default App;
