+++ src/Layouts/Layout.astro
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html lang='en'>
+  <head>
+    <meta charset='UTF-8' />
+    <meta http-equiv='X-UA-Compatible' content='IE=edge' />
+    <meta name='viewport' content='width=device-width, initial-scale=1.0' />
+    <link rel='preconnect' href='https://fonts.googleapis.com' />
+    <link rel='preconnect' href='https://fonts.gstatic.com' crossorigin />
+    <link
+      href='https://fonts.googleapis.com/css2?family=Inter:wght@100;200;300;400;500;600;700;800;900&family=Righteous&display=swap'
+      rel='stylesheet'
+    />
+    <link
+      href='https://fonts.googleapis.com/css2?family=Righteous&family=Russo+One&display=swap'
+      rel='stylesheet'
+    /> 
+    <style>
+      .container {
+        min-height: 100vh;
+        background-color: black;
+      }
+    </style>
+    <title>Fearless Tech</title>
+  </head>
+  <body>
+    <div class='container'>
+      <slot />
+    </div>
+  </body>
+</html>
