+++ src/components/Home/Hero/index.astro
@@ -1,0 +1,222 @@
+---
+import { hero_styles } from '../../../styles/colors';
+---
+
+<style
+  lang='scss'
+  define:vars={{
+    GREEN_HERO_TEXT: hero_styles.GREEN_HERO_TEXT,
+    PURPLE_HERO_TEXT: hero_styles.PURPLE_HERO_TEXT,
+    ft_purple: hero_styles.FEARLESSTECH_PURPLE,
+    ft_cyan: hero_styles.FEARLESSTECH_CYAN,
+  }}
+>
+  .hero {
+    min-height: 100vh;
+    background-position: center;
+
+    .border {
+      background-color: #d53a9d;
+      background-image: linear-gradient(
+        to left,
+        var(--GREEN_HERO_TEXT),
+        var(--PURPLE_HERO_TEXT)
+      );
+      position: absolute;
+      width: 100%;
+      height: 100%;
+      clip-path: polygon(
+        2% 2%,
+        2% 98%,
+        4% 97%,
+        4% 5%,
+        96% 5%,
+        96% 95%,
+        2% 95%,
+        2% 98%,
+        98% 98%,
+        98% 2%
+      );
+    }
+
+    * {
+      color: white;
+      box-sizing: border-box;
+      &::selection {
+        background-color: #771388d6;
+        color: white;
+      }
+    }
+
+    .content {
+      nav {
+        ul {
+          display: flex;
+          column-gap: 1rem;
+          justify-content: flex-start;
+          font-weight: 600;
+            border-bottom: 2px solid var(--ft_cyan);
+          width: max-content;
+          padding-bottom: 1rem;
+          li {
+            a {
+              &:hover {
+                color: var(--ft_cyan);
+              }
+            }
+          }
+        }
+      }
+      padding: 2rem 6rem;
+      display: flex;
+      flex-wrap: wrap;
+      justify-content: space-between;
+
+      .right-hero {
+        // right
+        min-height: 100vh;
+        display: flex;
+        width: 100%;
+        justify-content: flex-end;
+        position: absolute;
+        top: 0;
+        left: 0;
+        padding: 2rem 8rem;
+
+        .img-wrapper {
+          top: 0;
+          right: 0;
+          width: 70vh;
+          max-width: 600px;
+          position: relative;
+          display: flex;
+          align-items: center;
+          justify-content: center;
+          @keyframes float {
+            from {
+              transform: translateY(25px) scale(1.02);
+            }
+            to {
+              transform: translateY(45px);
+            }
+          }
+          &::before {
+            content: '';
+            background-color: rgba(19, 98, 177, 0.178);
+            box-shadow: 0 0 50px rgba(19, 98, 177, 0.456);
+            top: 30px;
+            width: 365px;
+            height: 365px;
+            border-radius: 100vmax; //brb
+            position: absolute;
+            animation: float alternate infinite backwards 2000ms ease-in-out;
+          }
+          img {
+            width: 85%;
+            animation: float alternate infinite 2000ms ease-in-out;
+            opacity: 95%;
+            transition: 200ms ease-in-out;
+            &:hover {
+              opacity: 100%;
+              backdrop-filter: brightness(1.19);
+            }
+          }
+        }
+      }
+      .left-hero {
+        // left
+        z-index: 50;
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        row-gap: 2rem;
+        min-height: 100vh;
+        margin-left: 8rem;
+        .last {
+          display: flex;
+          .community {
+            &::after {
+              transition: 0ms ease-in-out;
+              content: 'Community';
+              user-select: text;
+            }
+          }
+        }
+        h1 {
+          font-weight: 900;
+          font-size: 100px;
+          font-family: 'Righteous', cursive;
+          background: linear-gradient(var(--ft_purple), var(--ft_cyan));
+          -webkit-background-clip: text;
+          -webkit-text-fill-color: transparent;
+          letter-spacing: 5px;
+        }
+        .text {
+          display: flex;
+          flex-direction: column;
+          row-gap: 0.6rem;
+          
+          h4 {
+            display: block;
+            font-weight: 300;
+            font-size: 54px;
+            color: var(--GREEN_HERO_TEXT);
+            margin-left: 3rem;
+            &.purple-text {
+              color: var(--PURPLE_HERO_TEXT);
+              font-weight: 600;
+            }
+          }
+          .blink {
+            width: 5px;
+            height: 50px;
+            background-color: white;
+            animation: blink 1s infinite;
+            @keyframes blink {
+              0% {
+                opacity: 0;
+              }
+              49% {
+                opacity: 0;
+              }
+              50% {
+                opacity: 1;
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+</style>
+
+<div class='hero'>
+  <div class='border'></div>
+  <div class='content'>
+    <div class='left-hero'>
+      <nav>
+        <ul>
+          <li><a href='#'>Members</a></li>
+          <li><a href='#'>Projects</a></li>
+          <li><a href='#'>Education</a></li>
+          <li><a href='#'>Social</a></li>
+        </ul>
+      </nav>
+      <div class='text'>
+        <h4>Your</h4>
+        <h4>Favorite</h4>
+        <h4 class='purple-text'>Open</h4>
+        <h4 class='purple-text'>Source</h4>
+        <div class='last'>
+          <h4 class='community'></h4><div class='blink'></div>
+        </div>
+      </div>
+      <h1>FearlessTech</h1>
+    </div>
+    <div class='right-hero'>
+      <div class='img-wrapper'>
+        <img src='public/images/hero_image.png' alt='Wolf Silhouette' />
+      </div>
+    </div>
+  </div>
+</div>
