+++ src/components/Home/Intro/index.astro
@@ -1,0 +1,130 @@
+---
+import { intro_styles } from '../../../styles/colors';
+---
+
+<style
+  lang='scss'
+  define:vars={{
+    TEXT_CYAN: intro_styles.TEXT_CYAN,
+    TEXT_LT_CYAN: intro_styles.TEXT_LT_CYAN,
+    SKILLS_PURPLE: intro_styles.SKILLS_PURPLE,
+  }}
+>
+  .intro {
+    display: flex;
+    height: max(100vh, 500px);
+    min-height: 100vh;
+    max-width: 100%;
+    overflow: hidden;
+    flex-wrap: wrap;
+    justify-content: space-between;
+    * {
+      color: white;
+    }
+    .left {
+      flex: 1 1 250px;
+      display: flex;
+      flex-direction: column;
+      row-gap: 2rem;
+      justify-content: center;
+      padding-left: 2em;
+      h1 {
+        color: var(--SKILLS_PURPLE);
+        font-size: 26px;
+        font-weight: 300;
+        &:hover {
+          color: var(--TEXT_CYAN);
+          cursor: ;
+        }
+      }
+      h4 {
+        color: var(--SKILLS_PURPLE);
+        font-size: 22px;
+        font-weight: 200;
+      }
+    }
+
+    .center {
+      flex: 3 1 250px;
+      display: flex;
+      flex-direction: column;
+      row-gap: 2rem;
+      padding-top: 5rem;
+      h2 {
+        font-size: 50px;
+        font-weight: 400;
+        margin-left: 1rem;
+        line-height: 3.5rem;
+        .pop {
+          font-size: 55px;
+          font-weight: 600;
+          color: var(--TEXT_LT_CYAN);
+        }
+      }
+
+      .border {
+        border-width: 2px;
+        border-style: solid;
+        border-image: linear-gradient(to right, #f61cc7, #06626a) 1;
+        max-width: 100%;
+        padding: 2rem;
+        p {
+          font-size: 30px;
+          color: var(--TEXT_CYAN);
+        }
+      }
+    }
+
+    .right {
+      flex: 1 1 250px;
+      .img-wrapper {
+        display: flex;
+        justify-content: flex-end;
+        height: 700px;
+        min-height: 100%;
+        img {
+          height: 100%;
+        }
+      }
+    }
+  }
+</style>
+<div class='intro'>
+  <div class='left'>
+    <h1>UI/UX <br /> DESIGN</h1>
+    <h1>PROGRAMMING</h1>
+    <h1>DEVELOPMENT</h1>
+    <h1>DEPLOYMENT</h1>
+    <h1>APP TESTING</h1>
+    <h1>CYBER SECURITY</h1>
+    <h1>GRAPHIC ARTS</h1>
+    <h4>
+      Bring your cause,
+      <br />
+      connect with people.
+    </h4>
+  </div>
+
+  <div class='center'>
+    <h2>
+      We are an
+      <br />
+      <span class='pop'> OPEN TO ALL,</span>
+      <br />
+      Open Source
+      <br />
+      Community.
+    </h2>
+    <div class='border'>
+      <p>
+        We hope to bring together Programmers, Developers, UI/UX designers,
+        testers, and beyond to participate in creating social impact products.
+      </p>
+    </div>
+  </div>
+  <div class='right'>
+    <div class='img-wrapper'>
+      <img src='/images/blue_wolf.png' alt='Blue Wolf' />
+    </div>
+  </div>
+</div>
