--- src/components/Home/index.js
@@ -1,12 +1,0 @@
-import React from "react";
-import Hero from "./Hero";
-
-function Home() {
-	return (
-		<div>
-			<Hero />
-		</div>
-	);
-};
-
-export default Home;
