--- src/components/Pre.js
@@ -1,6 +1,0 @@
-import React from "react";
-function Pre(props) {
-  return <div id={props.load ? "preloader" : "preloader-none"}></div>;
-}
-
-export default Pre;
