--- src/components/colors.js
@@ -1,18 +1,0 @@
-export default {
-    TEXT_CYAN: '#00A5B5',
-    TEXT_LT_CYAN: '#0AA4B2', // skills hover
-    TEXT_PINK: '#F80FC6',
-    TEXT_LT_PINK: '#FA31CF',
-    TEXT_GREEN: '#159B22',
-    TEXT_LT_GREEN: '#08C91C',
-    TEXT_PURPLE: '#9F54E6',
-    TEXT_LT_PURPLE: '#8617EF',
-    SKILLS_PURPLE: '#DA00EE', 
-  };
-  
-  export const hero_styles = {
-    GREEN_HERO_TEXT: '#35FF49',
-    PURPLE_HERO_TEXT: '#9F54E6',
-    FEARLESSTECH_PURPLE: '#B400C5', // background: linear-gradient(180deg, #B400C5 0%, #0BAFBE 100%);
-    FEARLESSTECH_CYAN: '#0BAFBE',
-  }
