+++ src/pages/Home/Home.astro
@@ -1,0 +1,51 @@
+---
+import Hero from '../../components/Home/Hero/index.astro';
+import Intro from '../../components/Home/Intro/index.astro';
+import Mission from '../../components/Home/Mission/index.astro';
+import Join from '../../components/Home/Join/index.astro';
+import Pitch from '../../components/Home/Pitch/index.astro';
+---
+
+<style lang='scss'>
+  .bar {
+    width: 100%;
+    padding: 2rem; 
+    background-color: #000000;
+    z-index: 50;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    color: #c0c0c0;
+    font-weight: 100;
+    .text {
+      &::selection {
+        background-color: #721d46;
+        color: #ffffff;
+      }
+    }
+    &.has-text {
+      scroll-snap-align: center;
+      padding-top: 5rem;
+      flex-direction: column;
+      row-gap: 3rem;
+      .media {
+      }
+      span {
+        &::selection {
+          background-color: #721d46;
+          color: #ffffff;
+        }
+      }
+    }
+  }
+</style>
+<Hero />
+<Intro />
+<Mission />
+<Join />
+<Pitch />
+
+
+<div class='bar has-text'>
+  <span>© 2023 - FearlessTech - All rights reserved.</span>
+</div>
