+++ src/pages/index.astro
@@ -1,0 +1,9 @@
+---
+import '../styles/index.scss';
+import Layout from '../Layouts/Layout.astro';
+import Home from './Home/Home.astro';
+---
+
+<Layout>
+  <Home />
+</Layout>
