+++ todo.txt
@@ -1,0 +1,1 @@
+hero-text to be thinner and smaller
